requirements-python3.txt
backports-abc==0.5
certifi==2021.5.30
chardet==4.0.0
enum34==1.1.10
future==0.18.2
h5py==2.10.0
idna==2.10
jsonpatch==1.32
jsonpointer==2.1
numpy==1.16.6
Pillow==6.2.2
pybind11==2.6.2
pyzmq==19.0.2
qpth==0.0.15
requests==2.25.1
scipy==1.2.3
singledispatch==3.6.2
six==1.16.0
torchnet==0.0.4
tornado==6.1
tqdm==4.61.1
urllib3==1.26.6
visdom==0.1.8.9
websocket-client==0.59.0
jupyterlab
robustness